pyproject.toml
[project]
name = "gmail-mcp"
version = "0.1.0"
description = "MCP server for Gmail API access"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Henry Chien" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
]

[project.scripts]
gmail-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
